requirements.txt
Django==5.2
django-filter==25.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
pillow==11.3.0
python-dotenv==1.1.1
django-cors-headers==4.9.0
psycopg2-binary== 2.9.10
requests==2.32.5
dj-database-url==1.2.0
django-nested-admin==4.1.4
django-ckeditor-5==0.2.18
pytest==8.4.2
reportlab==4.4.4
pillow-heif==1.1.1




